[Refactor] Move `@contextmanager` helpers to `spec_decode/utils.py`
Move `patch_tensor_parallel_group`, `_maybe_eager_context`, and `_disable_flash_comm_v1_context` out of `llm_base_proposer.py` into `vllm_ascend/spec_decode/utils.py` (definitions and comments verbatim). `llm_base_proposer.py` now imports them from utils and drops the imports they used exclusively (`contextmanager`, `replace`, `_ps`, `CompilationMode`). UT monkeypatch target follows the move (`utils.get_forward_context`).
